requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
qdrant-client>=1.6.0
faiss-cpu>=1.7.4
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
plotly>=5.17.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
matplotlib>=3.7.0
requests>=2.31.0
pydantic>=2.0.0

